started incorporating current schedule
