+++ screens/Schedule.tsx
@@ -1,159 +1,98 @@
 import { StyleSheet, ScrollView, Button, ActivityIndicator, TextInput, KeyboardAvoidingView, Platform } from 'react-native';
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
+import { RootTabScreenProps } from '../types';
 
-import { Text, View, useThemeColor} from '../components/Themed';
-import DropDown from '../components/DropDown';
-import { Courses } from '../components/Course';
+import { Text, View, useThemeColor } from '../components/Themed';
 
-export default function Schedule() {
+export default function Schedule({ navigation }: RootTabScreenProps<'Home'>) {
   const color = useThemeColor({ light: 'black', dark: 'white' }, 'text');
-
   const [isLoading, setLoading] = useState(true);
-  const [data, setData] = useState({
-    'categories': [''],
-    'departments': [''],
-    'progLevels': [''],
-    'terms': ['']
+  const [user, setUser] = useState({
+    name: "",
+    legalName: "",
+    ufid: "",
+    apps: [],
+    email: "",
+    switchUser: false
   });
-  const [category, setCategory] = useState([]);
-  const [department, setDepartment] = useState([]);
-  const [term, setTerm] = useState([]);
-  const [progLevel, setProgLevel] = useState([]);
-  const [isSearching, setSearching] = useState(true)
-  const [results, setResults] = useState({
-    'COURSES': [],
-    'LASTCONTROLNUMBER': 0,
-    "RETRIEVEDROWS": 0,
-    "TOTALROWS": 0,
-  })
-  const [failed, setFailed] = useState(false)
-  const [title, setTitle] = useState('');
-  const [code, setCode] = useState('');
-  const [instructor, setInstructor] = useState('');
-  const [isTermsOpen, setTOpen] = useState(false);
-  const [isCatOpen, setCOpen] = useState(false);
-  const [isProgOpen, setPOpen] = useState(false);
-  const [isDepOpen, setDOpen] = useState(false);
 
   useEffect(() => {
-    axios.get("https://one.uf.edu/apix/soc/filters").then((res) => {
-      console.log(res);
-      setData(res.data);
-      setLoading(false);
-    }).catch((error) => {
-      console.log(error);
-    })
-  }, []);
-
-  function search() {
-    var url = "https://one.uf.edu/apix/soc/schedule?"
-    url = url + "ai=false&auf=false&category=" + category + "&class-num=&course-code=" + code + "&course-title=" + title + "&cred-srch=&credits=&day-f=&day-m=&day-r=&day-s=&day-t=&day-w=&dept=" + department;
-    url = url + "&eep=&fitsSchedule=false&ge=&ge-b=&ge-c=&ge-d=&ge-h=&ge-m=&ge-n=&ge-p=&ge-s=&instructor=" + instructor + "&last-control-number=0&level-max=&level-min=&no-open-seats=false&online-a=&online-c=&online-h=&online-p=&period-b=&period-e=&prog-level=" + progLevel;
-    url = url + "&qst-1=&qst-2=&qst-3=&quest=false&term=" + term + "&wr-2000=&wr-4000=&wr-6000=&writing=false&var-cred=&hons=false";
-    setLoading(true);
-    axios.get(url).then((res) => {
-      console.log(res);
-      setResults(res.data[0]);
-      setSearching(false);
-      setLoading(false);
-    }).catch((error) => {
-      setFailed(true);
-      setLoading(false);
-      console.log(error);
-
-    });
-  }
-
-  function returnToSearch() {
-    setSearching(true);
-    setFailed(false);
-  }
+    setLoading(false);
+  })
   return (
-    <View style={styles.container}>
-      {failed ? <View style={{ position: 'absolute', left: 0, top: '10%' }}>
-        <Button title="<Return to Search" onPress={returnToSearch} />
-        <Text style={[styles.title, { alignSelf: 'center' }]}>No results found</Text>
-      </View> :
-        isLoading ? <View style={{ alignSelf: 'center', alignContent: 'center', alignItems: 'center' }}><ActivityIndicator size={'large'} color={'blue'} /></View> :
-          <KeyboardAvoidingView style={styles.container} behavior={Platform.OS === 'ios' ? 'padding' : 'height'}>
-            {isSearching ?
-              <ScrollView style={styles.scrollView} contentContainerStyle={styles.contentContainer}>
-                <View style={[styles.container, {height: 120, marginBottom: isTermsOpen ? 200 : 0}]}>
-                  <Text style={styles.title}>Terms:</Text>
-                  <DropDown selectedValue={term} setSelectedValue={setTerm} items={data['terms']} open={isTermsOpen} setOpen={setTOpen} zIndex={1000}/>
-                </View>
-                <View style={[styles.container, {height: 120, marginBottom: isCatOpen ? 200 : 0}]}>
-                  <Text style={styles.title}>Categories:</Text>
-                  <DropDown selectedValue={category} setSelectedValue={setCategory} items={data['categories']} open={isCatOpen} setOpen={setCOpen}zIndex={2000}/>
-                </View>
+    isLoading ? <View style={{ alignSelf: 'center', alignContent: 'center', alignItems: 'center' }}><ActivityIndicator size={'large'} color={'blue'} /></View> :
+      <ScrollView >
+        <View style={{ alignItems: "center", flex: 1, paddingTop: 60, paddingBottom: 60 }}>
+          <Button title="Go to schedule of courses >" onPress={() => {
+            navigation.navigate("SOC")
+          }} />
+          <View style={styles.separator} />
 
-                <View style={[styles.container, {height: 120, marginBottom: isProgOpen ? 200 : 0}]}>
-                  <Text style={styles.title}>Program Levels:</Text>
-                  <DropDown selectedValue={progLevel} setSelectedValue={setProgLevel} items={['', ...data['progLevels']]} open={isProgOpen} setOpen={setPOpen} zIndex={3000}/>
-                </View>
-                <View style={[styles.container, {height: 120, marginBottom: isDepOpen ? 200 : 0}]}>
-                  <Text style={styles.title}>Departments:</Text>
-                  <DropDown selectedValue={department} setSelectedValue={setDepartment} items={['', ...data['departments']]} open={isDepOpen} setOpen={setDOpen} zIndex={4000}/>
-                </View>
-                <View style={[styles.container, {padding: '5%'}]}>
-                  <Text style={styles.title}>Course Title:</Text>
-                  <TextInput style={[styles.input, {color}]} onChangeText={setTitle} value={title} />
-                </View>
-                <View style={[styles.container, {padding: '5%'}]}>
-                  <Text style={styles.title}>Course Code:</Text>
-                  <TextInput style={[styles.input, {color}]} onChangeText={setCode} value={code} />
-                </View>
-                <View style={[styles.container, {padding: '5%'}]}>
-                  <Text style={styles.title}>Instructor Last Name:</Text>
-                  <TextInput style={[styles.input, {color}]} onChangeText={setInstructor} value={instructor} />
-                </View>
-                <Button title='Search' onPress={search} />
+          <View style={styles.personal_info}>
+            <Text style={styles.title}>{user.name}</Text>
 
-              </ScrollView> :
-              <View style={styles.container}>
-                <Courses courses={results['COURSES']} />
-                <View style={{ position: 'absolute', left: '0%', top: '10%' }}>
-                  <Button title="<Return to Search" onPress={returnToSearch} />
-                </View>
-              </View>
-            }
-          </KeyboardAvoidingView>
+          </View>
 
-      }
-    </View>
-  );
+        </View>
+
+      </ScrollView>
+
+  )
 }
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
+  personal_info: {
+    justifyContent: "flex-start",
+    alignItems: "center",
+    borderWidth: 3,
+    borderColor: "#285697",
+    marginTop: 1,
+    borderRadius: 15,
+    paddingBottom: 1,
+    paddingTop: 1,
+    width: '90%'
   },
+
   title: {
-    textAlign: 'center',
-    fontSize: 15,
+    fontSize: 20,
     fontWeight: 'bold',
+    alignSelf: 'center'
   },
   separator: {
     marginVertical: 30,
-    height: 1,
     width: '80%',
+    color: '#285697',
+    borderBottomColor: 'black',
+    borderBottomWidth: 1,
   },
-  input: {
-    height: 40,
-    borderWidth: 1,
-    padding: 10,
-    minWidth: '90%'
+  body: {
+    textAlign: 'center',
+    fontSize: 17,
   },
-  scrollView: {
+
+  subsection: {
     alignSelf: 'center',
-    alignContent: 'center',
+    width: '90%',
+    borderColor: 'grey',
+    borderWidth: 2,
+    borderRadius: 15,
+    marginTop: 10,
+    marginBottom: 10,
+    padding: 5,
+    paddingBottom: 15,
   },
-  contentContainer: {
-    justifyContent: 'center',
-    alignItems: 'center',
-    paddingTop: 80
+  section: {
+    alignSelf: 'center',
+    width: '90%',
+    borderColor: '#285697',
+    borderWidth: 2,
+    borderRadius: 15,
+    marginTop: 10,
+    marginBottom: 10,
+    padding: 5,
+    paddingBottom: 15,
+    alignItems: 'center'
   }
+
 });
